+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Added
+
+- Added initial GraphRAG implementation.
+- Added support for OpenCode in `init` cli command.
+- Added option to start embedding at the end of `init` cli command.
+
 ## 0.1.1
 
 - Fix issue with module not found error for `openai` when using Ollama only.